Colliders Fixed on All Ojbects
All of the main objects in the game have had their colliders setup or fixed. The player, enemies, and base all have a physical collider to provide them with ways to interact with the environment, and a "hitbox" collider which is a trigger that allows events to occurs when objects collide with each other.