Created Resource Scripts
Created basic resource scripts and update older scripts to work with a resource system.
There is a resource prefab.
The player can pick up the resource and deposit it at a base location in a dropoff trigger collider.
The base has a dropoff trigger collider associated with it.
The base has a resource stockpile which holds the current value of resource.
The base changes scaling based on the stockpile of resource within it.
The base has a resource win condition determined by bringing back a certain amount of resource to the base.
Added a scene called Alpha_NewGame where current single level was being focused on.
Changed some gameManager scripting so that at least the one scene containing the new current level and the win screen would work together.
Updated the MainMenuScript for the same reason.