Got Radar Working With Enemies
The Radar Scripts work for tracking enemies now
SpawnManager uses reference to RadarDisplay script on player so that it has access to the lists so it can add the radar objects from the enemy prefabs.
The enemy prefabs needed radar objects added to them as children.
EnemyCollision script updated so that before enemies are destroyed for any reason, they can reference the RadarDisplay script and remove their corresponding radar objects from the lists.